refactor: split instance.go by concern (SRD-040 M1)
Mechanical file split of the 1661-line instance.go — the one file in
internal/instance violating the one-entity-per-file rule — into
one-concern-per-file units per SRD-040 §3.3:

- lifecycle.go     — State enum + String, State/setState/LastErr/Done,
                     Run/Cancel/Terminate, settleFinalState
- correlation.go   — the conversation-key cluster (SRD-017): associate/
                     values/held, ProcessEvent (EventProcessor),
                     CorrelationKeys, validateAndAssociate, extendReceivers
- loop.go          — the single-writer event loop: loop, dropLoopState,
                     applyEvent + appliers, dispatchToParked, join
                     arbitration (recheck*/fireOrJoin), loop helpers
- projection.go    — addToSnap, GetTokens, TokenHistory
- runtimevars.go   — the runtime-var name consts, DataReader,
                     RuntimeVar/RuntimeVarNames
- eventproducer.go — RegisterEvent/UnregisterEvent/PropagateEvent,
                     InstanceID, EventProducer, interface assertions

instance.go keeps the Instance struct, New/NewFromEvent + options,
createTracks/seedParallelStart, and emit (374 lines).

Pure moves: zero signature or body changes; doc comments moved attached
to their declarations; per-file imports minimized (the wtasks alias and
its rationale comment now live only in loop.go, which uses it). Package
doc stays in instance.go.

Coverage top-up (the split re-exposed pre-existing uncovered lines to the
diff-coverage gate — moved lines count as changed): new unit tests for the
defensive branches in the moved code — RegisterEvent without a parent
producer, PropagateEvent delegation + parent-error wrap, the
EventProducer self-accessor, Run(nil ctx), a zero-track loop settling
Completed, onWaiting's stopping drop, fireOrJoin's unknown survivor
(eventproducer_test.go, lifecycle_test.go, loop_test.go). RuntimeVar's
three constructor-error wraps stay uncovered: they trigger only if
NewItemDefinition/NewItemAwareElement/NewParameter fail on a value the
switch itself constructs — unreachable by a caller.

Verified: go build, go test -race (instance + thresher), go vet,
golangci-lint 0 issues, gofmt clean; make ci diff-coverage gate PASS.
